Add files via upload
The SUM function adds values. You can add individual values, cell references or ranges or a mix of all three.  
The TEXT function lets you change the way a number appears by applying formatting to it with format codes.
The subtotal function in Excel is a formula that calculates a range of values based on a specified operation. It appears as =SUBTOTAL(function_num, ref1, [ref2],...), where function_num refers to the operation you want the formula to perform.